Add Vagara icons, main menu
