+++ Assets/Prefabs/Racer.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 14e107d56b245b14d84cee4df158f0f9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
