+++ Assets/Scripts/DuskEchoWorld.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7fae7fa645bc8848858df980cfd57f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
