+++ Assets/Textures/Old Guy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6cbd042a18ccd824d8ee4847d56ebcbf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
