+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,6 +9,9 @@ EditorBuildSettings:
     path: Assets/Scenes/SampleScene.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
   - enabled: 1
+    path: Assets/Scenes/main-menu.unity
+    guid: 9dbc24e48c455764c8d3c5c463107696
+  - enabled: 1
     path: Assets/Scenes/1-1.unity
     guid: 56439809156a4684697a08af6a7dfe50
   - enabled: 0
